'task3'

+++ src/model/integ/currentResponse.go
@@ -18,7 +18,7 @@ type RsCurrent struct {
 
 type RsCurrentWind struct {
 	Speed float64 `json:"speed"`
-	Deg   int     `json:"deg"`
+	Deg   float64     `json:"deg"`
 }
 
 type RsCurrentClouds struct {
